requirements.txt
hachoir
numpy
Pillow

#pyrogram==1.0.7

#pyrogram==1.1.13
pyrogram

requests
tgcrypto
sqlalchemy==1.3.23
psycopg2-binary
feedparser
#speedtest-cli
